Refactor: Migrate /clear` to the new SlashCommand interface
